+++ providers/databricks/pyproject.toml
@@ -77,6 +77,9 @@ dependencies = [
 "azure-identity" = [
     "azure-identity>=1.3.1",
 ]
+"fab" = [
+    "apache-airflow-providers-fab"
+]
 
 # The dependency groups should be modified in place in the generated file
 # Any change in the dependencies is preserved when the file is regenerated
