+++ providers/edge/pyproject.toml
@@ -62,6 +62,13 @@ dependencies = [
     "retryhttp>=1.2.0,!=1.3.0",
 ]
 
+# The optional dependencies should be modified in place in the generated file
+# Any change in the dependencies is preserved when the file is regenerated
+[project.optional-dependencies]
+"fab" = [
+    "apache-airflow-providers-fab"
+]
+
 [project.urls]
 "Documentation" = "https://airflow.apache.org/docs/apache-airflow-providers-edge/0.20.0pre0"
 "Changelog" = "https://airflow.apache.org/docs/apache-airflow-providers-edge/0.20.0pre0/changelog.html"
