+++ providers/edge/src/airflow/providers/edge/get_provider_info.py
@@ -100,4 +100,5 @@ def get_provider_info():
             }
         },
         "dependencies": ["apache-airflow>=2.10.0", "pydantic>=2.10.2", "retryhttp>=1.2.0,!=1.3.0"],
+        "optional-dependencies": {"fab": ["apache-airflow-providers-fab"]},
     }
